move waiters, fix linty stuff
